skeleton for tof array
