+++ electronics_abstract_parts/__init__.py
@@ -28,6 +28,8 @@ from .PassiveFilters import LowPassRc, DigitalLowPassRc, LowPassRcDac, PullupDel
 from .AbstractDiodes import BaseDiode, Diode, BaseDiodeStandardPinning, TableDiode
 from .AbstractDiodes import  ZenerDiode, TableZenerDiode, ProtectionZenerDiode
 from .AbstractLed import Led, RgbLedCommonAnode, LedColor, LedColorLike
+from .AbstractLed import IndicatorLed, VoltageIndicatorLed, IndicatorSinkRgbLed
+from .AbstractLed import IndicatorLedArray
 from .AbstractFets import Fet, FetStandardPinning, BaseTableFet, TableFet
 from .AbstractFets import SwitchFet, TableSwitchFet
 from .AbstractSolidStateRelay import SolidStateRelay, DigitalAnalogIsolatedSwitch
